Add SwimCoach and update dependency injection
A new SwimCoach class was added to the com.alexq66.springcoredemo.common package and SportConfig class to the config package. The bean dependency injection in the DemoController was also updated to use SwimCoach ('swimCoach') instead of the previous CricketCoach ('cricketCoach'). These changes demonstrated the usage of Spring's "Qualifier" annotation for specifying the required dependency.